requirements.txt
aniso8601==9.0.1
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
greenlet==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.14
Werkzeug==0.16.1
